First large set of changes to the core analyzer code for this new phase of work:
1. Improve the method by which references are represented in the
graph.  The original implementation added reference edges from
SourceClip objects to the referenced mob, slot, and individual
components. That was not exactly as originally as intended and added
unnecessarily difficulty when interpreting dumps of the data
structure.  Now it represents the references exactly as intended by
the AAF model, i.e. src clip -> mob -> slot -> component. It no longer
resolves components down to just the referenced region, however. I may
bring that back in the future.  This affected the base Visitor class
and the ResolveReferenceVisitor implementation.

2. Improved the dump capability to reflect changes above. It now has
the ability to dump the graph of contained objects starting with the
header (i.e. the AAF directly) or the combined reference and contained
objects starting with the top level composition.  The dump also shows
whether an object appears because it is contained by a parent or
referenced (via a soruce reference) by a parent object.  Dumps of the
data structure are now very easy to interpret - very important when
debugging the application.

3. Major restructuring of the TestResult class hieararchy. I found the
original implementation too complicated. Too many things that had to
be understood just to add a single result and to maintain the
aggregate results.  The new code retains the basic TopLevel,
PhaseLevel, TestLevel, and DetailLevel result hiearchy but simplifies
it such that (only) DetailLevel results report directly against a
single requirement. All other higher level results are the aggregate
of child results. A single call to consolidate results before they are
reported takes care of computing the aggregrates. The interface by
which detailed results are added to test level results is
simpler. There is now one call to add a detailed result (i.e. report a
result for a requirement).  The redundant "doc ref" attribute was
removed from the results. A document reference is associated with the
requirement and that's sufficient.  The relationship between test
results and there associated test is now simpler and the initial set
of covered requirements for a TestLevelTestResult object is now
automatically initialized from the associated test.  These changes
significantly reduced the size and complexity of the test result code
and remove a lot of repeated boiler plate code fragments that appeared
throughout code base.

4. Some minor rearrangement of certain files to better reflect their
role. Move reference resolution related files from
EditProtocolTestImpl back down to BaseTestImpl where it was to begin
with (it is unrelated to the EP per se). It must execute as part of
the load phase (in BaseTestImpl) to build and resolve the test graph
so that it can be dumped to verify correctness before moving onto the
more detailed EP tests. Renamed TestVisitor to DumpVisitor to better
reflect it (evolved) purpose.


5. Modified the default result output to include the reasons why a
test failed rather than just information about teh requirement that
failed.

6. The changes in item 1 above resovled erroneous mob dependency chain
errors that were reported for certain files in the ep 2006 sample file
set.  See tracker item 1761537 "invalid derivation chain failure for
sample file".

7. Resolve tracker item 1771387 "invalid slot id failure". Added
REQ_EP_257 to cover this case.

8. Resolve tracker item 1777304 "Unique slot names enforced file wide
not per mob. "

Sorry for the large comment. This is the result of synchronizing
several weeks worth of work saved in my local source repository.